The entire project, basically
I've been working on these components for a while, but I was waiting for the other contributors before pushing them to GitHub in case something conflicted.

The implementations of FileReader and TodoList were not originally in my scope, but I wanted to make sure the project was done.

As for the workload balance, I feel our original plan was fair since I had more experience and time, but I honestly went a bit too far with the project on my end.

If I could do this project again, I would at least share my experiences with the other group members during the development process, and have them make some of the simpler components.